refactor(git): increase naming consistency
