+++ libs/util/git/src/index.ts
@@ -1,3 +1,3 @@
-export { GitClientService } from "./git.service";
+export { GitClientService } from "./git-client.service";
 export {} from "./git.constants";
 export * from "./types";
